i32
Former-commit-id: d3ed4396115f0111fb915247b8d23daa54658010

+++ app/cache/category_options.go
@@ -18,7 +18,7 @@ import (
 	"go2o/core/service/dps"
 )
 
-func readToCategoryDropList(mchId int64) []byte {
+func readToCategoryDropList(mchId int32) []byte {
 	categories := dps.SaleService.GetCategories(mchId)
 	buf := bytes.NewBuffer([]byte{})
 	var f iterator.WalkFunc = func(v1 interface{}, level int) {
@@ -37,7 +37,7 @@ func readToCategoryDropList(mchId int64) []byte {
 }
 
 // 获取销售分类下拉选项
-func GetDropOptionsOfSaleCategory(mchId int64) []byte {
+func GetDropOptionsOfSaleCategory(mchId int32) []byte {
 	return readToCategoryDropList(mchId)
 }
