+++ core/domain/domain.go
@@ -12,13 +12,13 @@ package domain
 //聚合根
 type IAggregateRoot interface {
 	// 获取聚合根编号
-	GetAggregateRootId() int64
+	GetAggregateRootId() int32
 }
 
 //领域对象
 type IDomain interface {
 	// 获取领域对象编号
-	GetDomainId() int64
+	GetDomainId() int32
 }
 
 //值对象
